// src/lib/enums.ts
export enum Technologies {
    Ruby = "Ruby",
    Rails = "Rails",
    React = "React",
    Node = "Node",
    Nest = "Nest",
    Express = "Express",
    PostgreSQL = "PostgreSQL",
    MySQL = "MySQL",
    MongoDB = "MongoDB",
    Redis = "Redis",
    Docker = "Docker",
    AWS = "AWS",
    LegalTech = "Legal Tech",
    SQLite = "SQLite",
    NATS = "NATS",
    Microservices = "Microservices",
    Prisma = "Prisma",
    Socketio = "Socket.io",
    PHP = "PHP",
    Laravel = "Laravel",
    FinancialLogic = "Financial Logic",
    Swagger = "Swagger",
    OpenAPI = "OpenAPI",
    RESTAPI = "REST API",
    GraphQL = "GraphQL",
    Git = "Git",
    WebSockets = "WebSockets",
    ScalableSystems = "Scalable Systems",
    CleanArchitecture = "Clean Architecture",
}

export enum ProgrammingLanguages {
    JavaScript = "JavaScript",
    TypeScript = "TypeScript",
    Ruby = "Ruby",
    Java = "Java",
    Python = "Python",
    PHP = "PHP",
    Laravel = "Laravel",
    Node = "Node",
    Nest = "Nest",
    React = "React",
    Express = "Express",
    Rails = "Rails",
    MySQL = "MySQL",
    PostgreSQL = "PostgreSQL",
    MongoDB = "MongoDB",
    Redis = "Redis",
    Docker = "Docker",
    AWS = "AWS",
}

export enum SoftSkills {
    ProblemSolving = "Problem Solving",
    TeamCollaboration = "Team Collaboration",
    Communication = "Communication",
    QuickLearner = "Quick Learner",
}